Skip reading request body when body processing is disabled.
We have encountered an issue when SetRequestBodyAccess is turned off for
the code would still read the body into Apache brigades and that would
later cause processing to fail because it did not expect any body data
to be present.
Besides, that would allocate twice as much memory for body handling for
no good reason.

With this fix, we only copy the body over from IIS into Apache
structures if we are going to process it.

Signed-off-by: Vladimir Krivopalov <vlkrivop@microsoft.com>
(cherry picked from commit 24783aacdf265789c6316e4ad8a2fcfd6cdebbe0)